// pc/src/services/apis/models/FileDownloadVo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 企业全生命周期
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FileDownloadVo
 */
export interface FileDownloadVo {
    /**
     * 
     * @type {string}
     * @memberof FileDownloadVo
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof FileDownloadVo
     */
    path: string;
}


/**
 * Check if a given object implements the FileDownloadVo interface.
 */
export function instanceOfFileDownloadVo(value: object): value is FileDownloadVo {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('path' in value) || value['path'] === undefined) return false;
    return true;
}

export function FileDownloadVoFromJSON(json: any): FileDownloadVo {
    return FileDownloadVoFromJSONTyped(json, false);
}

export function FileDownloadVoFromJSONTyped(json: any, ignoreDiscriminator: boolean): FileDownloadVo {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'path': json['path'],
    };
}

export function FileDownloadVoToJSON(json: any): FileDownloadVo {
    return FileDownloadVoToJSONTyped(json, false);
}

export function FileDownloadVoToJSONTyped(value?: FileDownloadVo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'path': value['path'],
    };
}